service-workers/service-worker/fetch-event-respond-with-body-loaded-in-chunk.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>respondWith with a response whose body is being loaded from the network by chunks</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
'use strict';

const WORKER = 'resources/fetch-event-respond-with-body-loaded-in-chunk-worker.js';
const SCOPE = 'resources/fetch-event-respond-with-body-loaded-in-chunk-iframe.html';

promise_test(async t => {
  const registration =
      await service_worker_unregister_and_register(t, WORKER, SCOPE);
  await wait_for_state(t, registration.installing, 'activated');
  const iframe = await with_iframe(SCOPE);
  t.add_cleanup(async() => {
    if (iframe)
      iframe.remove();
    if (registration)
      await registration.unregister();
  });

  const response = await iframe.contentWindow.fetch('body-in-chunk');
  const result = await response.text()
  assert_equals(result, 'TEST_TRICKLE\nTEST_TRICKLE\nTEST_TRICKLE\n' +
                'TEST_TRICKLE\n');
}, 'Respond by chunks with a Response being loaded');
</script>
